+++ clients/go/ahp/reducers.go
@@ -498,6 +498,13 @@ func ApplyActionToSession(state *ahptypes.SessionState, action ahptypes.StateAct
 	case *ahptypes.SessionActivityChangedAction:
 		state.Summary.Activity = a.Activity
 		return ReduceOutcomeApplied
+	case *ahptypes.SessionChangesetsChangedAction:
+		if a.Changesets == nil {
+			state.Changesets = nil
+		} else {
+			state.Changesets = append([]ahptypes.Changeset(nil), a.Changesets...)
+		}
+		return ReduceOutcomeApplied
 	case *ahptypes.SessionConfigChangedAction:
 		if state.Config == nil {
 			return ReduceOutcomeNoOp